backend/requirements.txt
fastapi
uvicorn
requests
pyyaml
python-multipart
langgraph
httpx
PyJWT
passlib[bcrypt]
apscheduler
reportlab
